+++ heroku/interface_text.json
@@ -2,7 +2,10 @@
     "title": "# Tonami",
     "instructions": "Please wait for instructions from the interviewer.  \n   \n Please note that your audio will be recorded.  \n By clicking Next, you agree to allowing your recordings to be shared anonymously with the team and trusted evaluators.",
     "calibration": "According to all known laws of aviation, there is no way a bee should be able to fly. Do you think bees care what humans think is impossible?",
-    "calibration_end": "Thank you for patience. You may now proceed to the next module",
+    "calibration_end": [
+        "Thank you for patience. You may now proceed to the next module",
+        "Calibration complete!"
+    ],
     "exercises": [
         {
             "character": "罚",
