+++ tonami/controller.py
@@ -54,7 +54,7 @@ def load_exercise(filename: str = 'wo3_MV2_MP3.mp3'):
 
     return fig
 
-def process_user_audio (figure, user_info: dict[user.User], filename:str, tone: int=None):
+def process_user_audio (figure, user_info: dict[user.User], filename:str, classifier:c.Classifier, tone: int=None):
     """
     Takes the user's info, user's track and the desired tone/word
 
@@ -70,8 +70,7 @@ def process_user_audio (figure, user_info: dict[user.User], filename:str, tone: 
 
     user_utterance = u.Utterance(filename=filename)
     user_pitch_contour, user_nans, features = user_utterance.pre_process(user_info)
-    
-    classifier = c.Classifier(4, 'svm')
+
     classified_tones, classified_probs = classifier.classify_tones(features)
 
     # use the same axis as the native speaker's pitch contour plot
